+++ CHANGELOG.md
@@ -1038,3 +1038,4 @@
 - [2026-05-10 20:19] perf(ui): apply RepaintBoundary on animated ui widgets
 - [2026-05-10 20:36] test(state): add unit tests for state business logic
 - [2026-05-10 20:57] refactor(analytics): decouple repository interface from concrete analytics service
+- [2026-05-10 21:21] test(navigation): implement widget integration tests for navigation